==> pronunciation-prediction/01_preprocessing.ipynb <==
import pandas as pd
import numpy as np
from jamotools import split_syllables, Vectorizationer, join_jamos, rules
from unicodedata import normalize

data = pd.read_csv("korean_phonetic_spelling.csv", sep="\t")
print(len(data))
data.head()

# test jamotools

store = data.iloc[0]["spelling"]
store

split_syllables(store)

# test with string with punctuation

store_pron = data.iloc[0]["pronunciation"]
store_pron

split_syllables(store_pron)

# add these representations to the DataFrame

data["split_spelling"] = data.spelling.apply(split_syllables)

data.head()

data["split_pronunciation"] = data.pronunciation.apply(split_syllables)

data.head()

# find the longest sequence for each column

data.split_spelling.str.len().max()

data.split_pronunciation.str.len().max()

data[data.split_spelling.str.len() == 16]

data[data.split_pronunciation.str.len() == 17]

# test Vectorizationer

vec = Vectorizationer(rule=rules.RULE_1, max_length=None)

vec.vectorize(store)

vec.vectorize(store_pron)

# check out how punctuation and whitespace is handled

vec.vectorize(' .,/()"*:-%')

vec.vectorize("ː")

# vectorizer appears to work, so perhaps I can use this instead of split_syllables()

data["vec_spelling"] = data.spelling.apply(vec.vectorize)

data["vec_pronunciation"] = data.pronunciation.apply(vec.vectorize)

data.head()

# and while there appears to be no way to convert a vector back to text, split_syllables() just produces a string
# which join_jamos() is able to reconstruct so I can probably just reverse the symbol_map contained in a Vectorizer
# object

store

split = split_syllables(store)
split

type(split)

join_jamos(split)

dir(vec)

vec.symbol_map

decoder = {v: k for k, v in vec.symbol_map.items()}

decoder

def unvectorize(vector):
    temp = [decoder[num] for num in vector]
    return "".join(temp)

test_vec = vec.vectorize(store)
test_vec

unvectorize(test_vec)

# test on the DataFrame

data["unvec_spelling"] = data.vec_spelling.apply(unvectorize)

data["unvec_pronunciation"] = data.vec_pronunciation.apply(unvectorize)

data.head()

len(data[data.spelling != data.unvec_spelling])

# columns don't appear to match for any of the rows despite visual equality

data.iloc[0].spelling

data.iloc[0].unvec_spelling

data.iloc[0].spelling == data.iloc[0].unvec_spelling

# there appear to be different unicode representations as these words are exactly the same

"가게" == "가게"

# confirm this via ord() which returns an integer representing the unicode code point

[ord(x) for x in data.iloc[0].spelling]

[ord(x) for x in data.iloc[0].unvec_spelling]

# unicodedata is a Python module. its normalize() method can handle situations such as these by converting
# code points to normal form.
# "NFD", normal form D, is also known as canonical decomposition and translates each character into its decomposed form.
# "NFC" first applies the canonical decomposition, then composes the pre-composed characters again

normalize("NFD", data.iloc[0].spelling) == normalize("NFD", data.iloc[0].unvec_spelling)

# unvec_spelling originally contained strings in "NFD"

[ord(x) for x in normalize("NFD", data.iloc[0].spelling)]

# spelling originally contained strings in "NFC"

[ord(x) for x in normalize("NFC", data.iloc[0].spelling)]

# alter the function to produces values in line with the original data

def unvectorize_norm(vector):
    temp_list = [decoder[num] for num in vector]
    temp_string = "".join(temp_list)
    return normalize("NFC", temp_string)

unvectorize_norm(test_vec)

[ord(x) for x in normalize("NFC", unvectorize_norm(test_vec))]

# re-test unvectorization, this time with unvectorize_norm

data = data.drop(["unvec_spelling", "unvec_pronunciation"], axis=1)
data.head()

data["unvec_spelling"] = data.vec_spelling.apply(unvectorize_norm)

data["unvec_pronunciation"] = data.vec_pronunciation.apply(unvectorize_norm)

len(data[data.spelling != data.unvec_spelling])

data

# <UNK> appears to be the only issue

len(data[data.pronunciation != data.unvec_pronunciation])

len(data[data.unvec_pronunciation.str.contains("<UNK>")])

data[data.unvec_pronunciation.str.contains("<UNK>")].head()

# this punctuation isn't a colon but actually two triangles facing each other creating the shape of an hourglass.
# (see below for a visual; the code point is U+2d0)
# the IPA (international phonetic alphabet) uses this symbol to express length (how long a sound is pronounced)

https://en.wikipedia.org/wiki/Length_(phonetics)

hex(ord("ː"))

%%html 

<h1>ː</h1>

# convert this to a common colon so Vectorizationer() can handle it properly

data.pronunciation = data.pronunciation.str.replace("ː", ":")
data.split_pronunciation = data.split_pronunciation.str.replace("ː", ":")

# vectorize the pronunciation column again

data["vec_pronunciation"] = data.pronunciation.apply(vec.vectorize)

# unvectorize pronunciation and compare it to the pronunciation column

data["unvec_pronunciation"] = data.vec_pronunciation.apply(unvectorize_norm)

len(data[data.pronunciation != data.unvec_pronunciation])

data.head()

# this methods appears to be successful for converting back and forth between Korean and the vectors

# NOTE: 113 is the colon in the reversed symbol_map dictionary

decoder[113]

# to make sure the dimensions of each row are the same, pad the end of each vector with zeros until it is as long as
# the longest vector. this was calculated earlier but here again for convenience

data.split_spelling.str.len().max()

data.split_pronunciation.str.len().max()

# pad each vector in both vec_spelling and vec_pronunciation with zeros until it reaches 17 digits.
# though vec_spelling technically only needs to be padded until 16, making all vectors the same length will make
# the problem easier to solve

def pad_with_zeros(vector, num_zeros):
    zeros_to_pad = num_zeros - len(vector)
    # default value for "constant" mode is zero.
    # pad the array with 0 zeros on the left and "zeros_to_pad" zeros on the right
    return np.pad(vector, (0, zeros_to_pad), "constant")

# test

pad_with_zeros(data.iloc[0].vec_spelling, 17)

len(pad_with_zeros(data.iloc[0].vec_spelling, 17))

# apply
# NOTE: passing args to apply required a comma after the integer

data["vec_spelling_pad"] = data.vec_spelling.apply(pad_with_zeros, args=(17,))

data["vec_pronunciation_pad"] = data.vec_pronunciation.apply(pad_with_zeros, args=(17,))

data.head()

# convert these padded vectors to DataFrames

vec_spelling_df = pd.DataFrame(data.vec_spelling_pad.to_list())
vec_spelling_df.head()

vec_pronunciation_df = pd.DataFrame(data.vec_pronunciation_pad.to_list())
vec_pronunciation_df.head()

# do some quick confirmation of process. 8170 was the subscript of the longest pronunciation (checked above), so
# confirm that that's true here

data[data.split_pronunciation.str.len() == 17]

vec_pronunciation_df.iloc[8170]

unvectorize_norm(vec_pronunciation_df.iloc[8170].to_list())

# write to file

vec_spelling_df.to_csv("vec_spelling.csv", index=False, sep="\t")

vec_pronunciation_df.to_csv("vec_pronunciation.csv", index=False, sep="\t")

# confirm content written correctly

confirm = pd.read_csv("vec_spelling.csv", sep="\t")
confirm.head()

len(confirm)

confirm = pd.read_csv("vec_pronunciation.csv", sep="\t")
confirm.head()

len(confirm)
