==> pronunciation-prediction/02_seq2seq_TEST.ipynb <==
import pandas as pd
import numpy as np
from keras.models import Sequential
from keras.layers import Embedding, Bidirectional, LSTM, RepeatVector, TimeDistributed, Dense
from keras.initializers import VarianceScaling, RandomUniform
from keras.utils.np_utils import to_categorical
from jamotools import Vectorizationer, rules
from unicodedata import normalize

spelling = pd.read_csv("vec_spelling.csv", sep="\t")
spelling.head()

# the three parameters for the "encoding" step: vocab_size, output_dim, and num_cols

# the input dimension/vocabulary size for the model will be the number of rows passed into the embedding layer for training

vocab_size = spelling.shape[0]
vocab_size

# the output dimension for the embedding layer is the number of columns/parameters

output_dim = spelling.shape[1]
output_dim

pronunciation = pd.read_csv("vec_pronunciation.csv", sep="\t")
pronunciation.head()

# As the decoder RNN's input, repeatedly provide with the last output of
# RNN for each time step. Repeat "number of columns" times as that's the maximum
# length of output of the model

num_cols = pronunciation.shape[1]
num_cols

# the parameter for the decoding step: onehot_cols

# one-hot encode the target so that all mistakes are treated the same

onehot_pronunciation = to_categorical(pronunciation)
onehot_pronunciation.shape

# inspect the one-hot encoding

pronunciation.iloc[0]

len(onehot_pronunciation[0])

len(onehot_pronunciation[0][0])

# the first value encoded here should match the first value from pronunciation.iloc[0]

onehot_pronunciation[0][0]

# the output layer should have as many columns as now exist in the one-hot encoded array

onehot_cols = onehot_pronunciation.shape[2]
onehot_cols

# NMT models can be divided into two parts: the encoder and the decoder

# the encoder:
# an embedding layer to create word vectors from the input language; an LSTM layer; a RepeatVector which should be repeated
# as many times as the length of the output.
# NOTE: the Bidirectional wrapper around the lSTM layer trains two LSTMs instead of one, with one of the layers provided
#       with reversed copies of the input sequences. this provides more context to the network and can result in faster
#       training and better learning.
# NOTE: mask_zero=True indicates that input value zero is a special "padding" value in the vocabulary that should be
#       masked out. if this is set to True, the vocabulary size should be increased by +1 as index zero cannot be used

# the decoder:
# an LSTM layer which returns sequences (returns the hidden state output at each time step). a TimeDistributed/Dense layer
# which adds one Dense node to each unit of the previous layer. 
# NOTE: using an lSTM layer which returns sequences and a TimeDistributed layer together like this is useful when comparing
#       an entire sequence instead of just a final result like in classification. this way, the loss function is computed
#       for each token

model = Sequential()
model.add(Embedding(vocab_size+1, output_dim, trainable=True, mask_zero=True, 
                    embeddings_initializer=RandomUniform(minval=-0.05, maxval=0.05, seed=42)))
model.add(Bidirectional(LSTM(128, recurrent_activation="hard_sigmoid", 
                             kernel_initializer=VarianceScaling(distribution="uniform", mode="fan_avg"))))
model.add(RepeatVector(num_cols))
model.add(LSTM(256, recurrent_activation="hard_sigmoid", return_sequences=True,
                             kernel_initializer=VarianceScaling(distribution="uniform", mode="fan_avg")))
model.add(TimeDistributed(Dense(onehot_cols, activation="softmax",
                               kernel_initializer=VarianceScaling(distribution="uniform", mode="fan_avg"))))

model.compile(loss="categorical_crossentropy", optimizer="adam", metrics=["accuracy"])

model.summary()

model.fit(spelling, onehot_pronunciation, epochs=10)

# test the model

# first give it something it's already been trained on

# NOTE: make sure to reshape input before predicting

spelling.iloc[0].shape

test_easy = spelling.iloc[0].values.reshape(1, spelling.iloc[0].shape[0])
test_easy.shape

test_easy

# use subscript [0] on output of predict as it returns an array of predictions. 
# (the actual shape for a single prediction is (1, 17, 114))

prediction = model.predict(test_easy)[0]

# the shape stored in prediction now is num_cols x onehot_cols.
# there are 17 arrays of length 114 as there are 17 predictions to be made and 114 possibilites to choose from.
# access the highest probability for each prediction using .argmax() with axis=1

prediction.argmax(axis=1)

# the model predicted the exact same pronunciation as spelling for this word

(test_easy == prediction.argmax(axis=1)).all()

# NOTE however, that the pronunciation is actually different

spelling.iloc[0] == pronunciation.iloc[0]

# test to see if the model simply learned to output the exact same thing it received

test_output = spelling.iloc[4].values.reshape(1, spelling.iloc[4].shape[0])
test_output

prediction = model.predict(test_output)[0]

prediction.argmax(axis=1)

# the output isn't exactly the same, and this one is actually right

(test_output == prediction.argmax(axis=1)).all()

(pronunciation.iloc[4].values == prediction.argmax(axis=1)).all()

# it certainly isn't producing jibberish, but not all the predictions were correct, even with
# the same data it trained on, so try with more training data

vec = Vectorizationer(rule=rules.RULE_1, max_length=None)

decoder = {v: k for k, v in vec.symbol_map.items()}

def unvectorize_norm_pad(vector):
    temp_list = [decoder[num] for num in vector if num != 0]
    temp_string = "".join(temp_list)
    return normalize("NFC", temp_string)

def unvectorize_norm_prediction(prediction):
    vector = [np.argmax(x) for x in prediction]
    temp_list = [decoder[num] for num in vector if num != 0]
    temp_string = "".join(temp_list)
    return normalize("NFC", temp_string)

# spelling

unvectorize_norm_pad(spelling.iloc[4])

# predicted pronunciation

unvectorize_norm_prediction(prediction)

# actual pronunciation

unvectorize_norm_pad(pronunciation.iloc[4])

unvectorize_norm_prediction(prediction) == unvectorize_norm_pad(pronunciation.iloc[4])
