==> pronunciation-prediction/03_more_preprocessing.ipynb <==
import re
import numpy as np
import pandas as pd
from jamotools import Vectorizationer, rules
from unicodedata import normalize

# specify additional null value "NONE"

data = pd.read_csv("../web-scraping/pronunciation_all.csv", sep="\t", na_values="NONE")
data.head()

data.info()

data[data.pronunciation.isnull()].head(15)

# drop rows with null values in pronunciation before masking to check for English and Chinese loan words

data_drop_null = data[data.pronunciation.notna()]
len(data_drop_null)

# check for Chinese characters

data_drop_null[data_drop_null.pronunciation.str.contains(u'[\u4e00-\u9fff]', regex=True)]

data_drop_null[data_drop_null.spelling.str.contains(u'[\u4e00-\u9fff]', regex=True)]

# check for English lowercase letters 

data_drop_null[data_drop_null.pronunciation.str.contains(u'[\u0061-\u007A]', regex=True)]

data_drop_null[data_drop_null.spelling.str.contains(u'[\u0061-\u007A]', regex=True)]

# check for English uppercase letters

data_drop_null[data_drop_null.pronunciation.str.contains(u'[\u0041-\u005A]', regex=True)]

data_drop_null[data_drop_null.spelling.str.contains(u'[\u0041-\u005A]', regex=True)]

# check for duplicates
# NOTE: duplicates are NOT checked on the subset of spelling or pronunciation alone as identical combinations of spelling
#       and pronunciation can be linked to multiple word senses (definitions). as these different senses have different 
#       distributions they were conceived to be legitimate, albeit duplicated, members of the training
#       set. for example, the multiple instances of "시대적"

data_drop_null[data_drop_null.spelling == "시대적"]

len(data_drop_null.drop_duplicates())

data_drop_null[data_drop_null.duplicated()].head(15)

# NOTE the duplication of row 2425 and 2422

data_drop_null[data_drop_null.entry_id == 15639]

data_drop_dups = data_drop_null.drop_duplicates()
len(data_drop_dups)

# handle pronunciations containing forward slashes ("/")

forward_slash = data_drop_dups[data_drop_dups.pronunciation.str.contains("/")]
len(forward_slash)

# define function for splitting these rows

def split_row(row, row_accumulator, target_column, separator):
    split_row = row[target_column].split(separator)
    if len(split_row) != 2:
        print(row)
    for s in split_row:
        new_row = row.to_dict()
        new_row[target_column] = s
        row_accumulator.append(new_row)

new_rows = []
forward_slash.apply(split_row, axis=1, args=(new_rows, "pronunciation", "/"))
temp_df = pd.DataFrame(new_rows)
len(temp_df)

temp_df.head()

no_slash = data_drop_dups[~data_drop_dups.pronunciation.str.contains("/")]
print(len(data_drop_dups) - len(forward_slash))
len(no_slash)

# recombine DataFrames and handle any duplicates that resulted from splitting on forward slashes

combined = pd.concat([no_slash, temp_df])
print(len(no_slash) + len(temp_df))
len(combined)

combined.head()

print(len(combined[combined.duplicated()]))
combined[combined.duplicated()]

combined[combined.entry_id == 23444]

combined_drop_dups = combined.drop_duplicates()
len(combined_drop_dups)

# replace all instances of the phonetic "length" symbol "ː" with an ordinary colon ":"

len(combined_drop_dups[combined_drop_dups.pronunciation.str.contains("ː")])

# NOTE: a false positive SettingWithCopy warning will be raised here initially as .str.replace() does indeed
#       return a copy. Using .loc to access the original DataFrame however ensures that the original is modified
#       in this assignment statement

combined_drop_dups.loc[:,"pronunciation"] = combined_drop_dups.pronunciation.str.replace("ː", ":")

len(combined_drop_dups[combined_drop_dups.pronunciation.str.contains("ː")])

# create the Vectorizationer() instance

vec = Vectorizationer(rule=rules.RULE_1, max_length=None)

# test it out

vec.vectorize("가게")

# vectorize the spelling and pronunciation columns

combined_drop_dups.loc[:,"vec_spelling"] = combined_drop_dups.spelling.apply(vec.vectorize)

combined_drop_dups.loc[:,"vec_pronunciation"] = combined_drop_dups.pronunciation.apply(vec.vectorize)

combined_drop_dups.head()

# find the longest string of letters in either the spelling or pronunciation columns

combined_drop_dups.vec_spelling.apply(len).max()

combined_drop_dups[combined_drop_dups.vec_spelling.apply(len) == 23]

# NOTE use of .loc here

combined_drop_dups.loc[65120].vec_spelling

combined_drop_dups.vec_pronunciation.apply(len).max()

combined_drop_dups[combined_drop_dups.vec_pronunciation.apply(len) == 23]

combined_drop_dups.loc[65120].vec_pronunciation

# create a decoder and write a function to unvectorize the spelling and pronunciation to double-check output before padding

decoder = {v: k for k, v in vec.symbol_map.items()}

def unvectorize_norm(vector):
    temp_list = [decoder[num] for num in vector]
    temp_string = "".join(temp_list)
    return normalize("NFC", temp_string)

combined_drop_dups.loc[:,"unvec_spelling"] = combined_drop_dups.vec_spelling.apply(unvectorize_norm)

combined_drop_dups.loc[:,"unvec_pronunciation"] = combined_drop_dups.vec_pronunciation.apply(unvectorize_norm)

len(combined_drop_dups[combined_drop_dups.pronunciation != combined_drop_dups.unvec_pronunciation])

combined_drop_dups[combined_drop_dups.pronunciation != combined_drop_dups.unvec_pronunciation]

# these 21 entries for the vowel and compound vowel entries seem to be the only issue for pronunciation

len(combined_drop_dups[combined_drop_dups.vec_pronunciation.apply(len) == 1])

# remove the vowel only entries

combined_clean = combined_drop_dups[~(combined_drop_dups.pronunciation != combined_drop_dups.unvec_pronunciation)]
print(len(combined_drop_dups) - len(combined_drop_dups[combined_drop_dups.vec_pronunciation.apply(len) == 1]))
len(combined_clean)

# check spelling now

len(combined_clean[combined_clean.spelling != combined_clean.unvec_spelling])

len(combined_clean[combined_clean.unvec_spelling.str.contains("<UNK>")])

# Korean has 14 consonants, plus 5 double consonants, and ㄴ, ㄹ, ㅁ, and ㅇ seemed to have been treated as special cases.
# So four entries each, except for the special cases which only have one entry

(14 + 5 - 4)* 4 + 4

# just drop these entries like the vowels

combined_clean[combined_clean.unvec_spelling.str.contains("<UNK>")].head(40)

combined_clean_final = combined_clean[~(combined_clean.spelling != combined_clean.unvec_spelling)]
print(len(combined_clean) - len(combined_clean[combined_clean.unvec_spelling.str.contains("<UNK>")]))
len(combined_clean_final)

# pad all vectors with zeroes to ensure uniform length

def pad_with_zeros(vector, num_zeros):
    zeros_to_pad = num_zeros - len(vector)
    # default value for "constant" mode is zero.
    # pad the array with 0 zeros on the left and "zeros_to_pad" zeros on the right
    return np.pad(vector, (0, zeros_to_pad), "constant")

combined_clean_final.loc[:,"vec_spelling_pad"] = combined_clean_final.vec_spelling.apply(pad_with_zeros, args=(23,))

combined_clean_final.loc[:,"vec_pronunciation_pad"] = combined_clean_final.vec_pronunciation.apply(pad_with_zeros, args=(23,))

combined_clean_final.head()

# final stats

# as the model is using individual letters rather than words, its input is approximately 700,000 pairs rather than
# 90,000 pairs

combined_clean_final.vec_spelling.apply(len).describe()

combined_clean_final.vec_pronunciation.apply(len).describe()

# convert the padded vectors to DataFrames

vec_spelling_all = pd.DataFrame(combined_clean_final.vec_spelling_pad.to_list())
print(len(vec_spelling_all))
vec_spelling_all.head()

vec_pronunciation_all = pd.DataFrame(combined_clean_final.vec_pronunciation_pad.to_list())
print(len(vec_pronunciation_all))
vec_pronunciation_all.head()

# write to file

vec_spelling_all.to_csv("vec_spelling_all.csv", index=False, sep="\t")

vec_pronunciation_all.to_csv("vec_pronunciation_all.csv", index=False, sep="\t")

# write to file data on entries used for training for use with validation

combined_clean_final[["entry_id", "word_id", "spelling", "pronunciation"]].to_csv("reference_all.csv", 
                                                                                  index=False, sep="\t")
