==> pronunciation-prediction/05_prep_for_validation_01.ipynb <==
# the website used for IPA transcription percent-encoded data using a non-standard method.
# a headless browser circumvented the need to decipher this method.

from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.keys import Keys
import os
import time
import pickle
from random import randint

chrome_options = Options()

chrome_options.add_argument("--headless")

chrome_driver = os.getcwd() + "\\chromedriver.exe"

driver = webdriver.Chrome(options=chrome_options, executable_path=chrome_driver)

input_data = {}
output_data = {"resyllabification": [],
               "unrelease": [],
               "nasal_assimilation": [],
               "rieul_assimilation": [],
               "tensification": [],
               "aspiration_and_hieut_weakening": [],
               "double_consonant_reduction": [],
               "palatalization": [],
               "place_assimilation": []
              }

input_data["resyllabification"] = [
    "한글은",
    "읽어요",
    "책을",
    "알았어요",
    "질문이",
    "있어요",
    "읽어",
    "들으세요",
    "맞았어요",
    "앉으세요",
    "천만에요",
    "책이",
    "없어요",
    "백화점에",
    "갔어요",
    "옷을",
    "받았어요",
    "꽃은",
    "잎이",
    "같아요",
    "갔어",
    "낮에",
    "빛이",
    "부엌에",
    "낚아요",
    "꽃이"
]

# (Cho et al., 2009, p. 34)

input_data["unrelease"] = [
    "꽃",
    "꽃도",
    "잎",
    "잎과",
    "옷",
    "옷도",
    "갔다",
    "갔지",
    "밤낮",
    "낮과",
    "빛",
    "빛조차",
    "부엌",
    "부엌바닥",
    "낚시",
    "낚다가",
    "꽃씨"
]

# (Cho et al., 2009, p.35)

input_data["nasal_assimilation"] = [
    "입만",
    "앞문",
    "없나요",
    "받는다",
    "끝나다",
    "몇년",
    "있는데",
    "일학년",
    "낳는다",
    "모르겠습니다",
    "한국말로",
    "합니까",
    "끝내겠습니다"
]

# (Cho et al., 2009, p. 36)

input_data["rieul_assimilation"] = [
    "칠년",
    "물약",
    "진리",
    "신라",
    "전라도",
    "달님",
    "팔년",
    "서울역",
    "길이름"
]

# (Cho et al., 2009, p. 36)

input_data["tensification"] = [
    "몇번",
    "학생",
    "몇과",
    "식당",
    "학교",
    "숙제",
    "꽃집",
    "책상",
    "몇시간",
    "뵙겠습니다",
    "없습니다",
    "여름방학",
    "길가",
    "강가",
    "봄비",
    "누구거예요",
    "마이클거예요"
]

# (Cho et al., 2009, p. 36-37)

input_data["aspiration_and_hieut_weakening"] = [
    "좋다",
    "닫히다",
    "좋고",
    "입학",
    "많다",
    "좋지않다",
    "책하고",
    "시작합시다",
    "어떻게",
    "꽃하고",
    "대답하세요",
    "좋아요",
    "말한다",
    "전화",
    "여름방학",
    "사랑한다",
    "안녕하세요",
    "많아요",
    "좋았어요",
    "잘했어요",
    "천천히",
    "말해",
    "감사합니다",
    "괜찮아요" 
]

# (Cho et al., 2009, p. 37-38)

input_data["double_consonant_reduction"] = [
    "값",
    "값도",
    "여덟개",
    "없고",
    "읽습니다",
    "없습니다",
    "괜찮습니다",
    "앉겠어요"
]

# (Cho et al., 2009, p. 38)

input_data["palatalization"] = [
    "닫혀요",
    "붙이다",
    "붙여요",
    "같이",
    "밑이"
]

# (Cho et al., 2009, p. 38-39)

input_data["place_assimilation"] = [
    "꽃병",
    "갔고",
    "한미",
    "빗물",
    "한강",
    "옷감",
    "신문",
    "옷빨래",
    "한국어",
    "젖병",
    "눈꺼풀",
    "닫고",
    "꽃무늬",
    "밭갈이",
    "신발"
]

# (Cho et al., 2009, p. 39)

for key in input_data:
    for word in input_data[key]:

        driver.get("http://pronunciation.cs.pusan.ac.kr/pronunc.htm")

        form = driver.find_element_by_name("text1")

        submit_button = driver.find_element_by_id("submit1")

        form.clear()
        form.send_keys(word)

        submit_button.click()

        result = driver.find_elements_by_class_name("td2")
        
        if result is None:
            print()
            print(f"problem with {key}: {word}:")
            print("result is None")
            print()
            continue

        if result[0].text != result[1].text:
            print()
            print(f"problem with {key}: {word}:")
            for item in result:
                print(item.text)
            print()
        else:
            output_data[key] += [[result[1].text, result[2].text, result[3].text]]

        time.sleep(randint(10, 20))

# manually handling problem cases

output_data["nasal_assimilation"] += [["몇년", "면년", "myʌn ɲyʌn"]]

output_data["nasal_assimilation"] += [["일학년", "일항년", "il haŋɲyʌn"]]

output_data["rieul_assimilation"] += [["칠년", "칠련", "tsʰil  lyʌn"]]

output_data["rieul_assimilation"] += [["팔년", "팔련", "pʰal lyʌn"]]

output_data["tensification"] += [["몇번", "멷뻔", "myʌt̚ p̕ʌn"]]

output_data["tensification"] += [["몇시간", "멷시깐", "myʌt̚ ɕik̕an"]]

output_data["tensification"] += [["누구거예요", "누구꺼예요", "nugu k̕ʌyeyo"]]

output_data["aspiration_and_hieut_weakening"] += [["좋지않다", "조치안타", "tsotsʰi antʰa"]]

output_data["double_consonant_reduction"] += [["여덟개", "여덜개", "yʌdʌl kɛ"]]

output_data["place_assimilation"] += [["옷빨래", "옫빨래", "ot̚ p̕allɛ"]]

output_data

f = open("validation_data_integrated_korean.p", "wb")

pickle.dump(output_data, f)

f.close()

# References

Cho, Y., Lee, H. S., Schulz, C., Sohn, H., & Sohn, S. (2009). Integrated Korean: Beginning 1 (Klear Textbooks in Korean Language). University of Hawaii Press.
