==> pronunciation-prediction/05_prep_for_validation_02.ipynb <==
import pickle

# hardcoded data here to preserve original IPA transcriptions and to
# handle phenomena produced across word boundaries (Busan National University tool splits
# input on word boundaries).

introducing_phonology = {}

introducing_phonology["nasalization"] = [
#     ["밥", "밥", "pap"],
#     ["밥사", "밥사", "pap sa"],
#     ["밥도둑이", "밥도두기", "pap totuki"],
    ["밥많이", "밤마니", "pam mani"],
    ["밥먹으면", "밤머기면", "pam mək-imyən"],
    ["밥너무", "밤너무", "pam nəmu"],
    ["밥넣고", "밤너코", "pam nəkʰo"],
    ["입어", "이버", "ipə"],
    ["입는다", "임는다", "imnɨnta"],
    ["굽어", "구버", "kupə"],
    ["굽는다", "굼는다", "kumnɨnta"],
    ["접어", "저버", "čəpə"],
    ["접는다", "점는다", "čəmnɨnta"],
    ["닫아", "다다다", "tatə"],
    ["닫는다", "단는다", "tannɨnta"],
    ["붙어", "부터", "putʰə"],
    ["붙는다", "분는다", "punnɨnta"],
    ["쫓아", "쪼차", "čočʰa"],
    ["쫓는다", "쫀는다", "čonnɨnta"],
    ["먹어", "머거", "məkə"],
    ["먹는다", "멍는다", "məŋnɨnta"],
    ["섞어", "서꺼", "səkʼə"],
    ["섞는다", "성는다", "səŋnɨnta"],
    ["닦아", "다까", "takʼa"],
    ["닦는다", "당는다", "taŋnɨnta"],
    ["죽어", "주거", "čukə"],
    ["죽는다", "중는다", "čuŋnɨnta"],
    ["익어", "이거", "ikə"],
    ["익는다", "잉는다", "iŋnɨnta"],
]

# (Odden, 2005, p. 81-83)

f = open("validation_data_introducing_phonology.p", "wb")

pickle.dump(introducing_phonology, f)

f.close()

# References

Odden, D. (2005). Introducing Phonology (Cambridge Introductions to Language and Linguistics). Cambridge University Press.
